requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
imbalanced-learn==0.13.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
narwhals==1.25.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.5.0
tzdata==2025.1
